RWKV S90: whole-day answer search is test-only; S90 tuple annotation
The unrounded search replaced intervals_for_answer_curves in production,
so it and its two helpers are #[cfg(test)] now (the tests' oracle that the
unrounded values round up to the whole-day ones); clippy -D warnings on
all targets flagged them as dead. predict_current_intervals_many's return
annotation carries a float S90 (mypy).

+++ qt/aqt/rwkv_srs_benchmark.py
@@ -863,7 +863,7 @@ class _RustRwkvRuntime:
     def predict_current_intervals_many_from_warm_up(
         self,
         review_inputs: Sequence[RwkvReviewInput],
-    ) -> Sequence[tuple[float, int | None, int | None]]:
+    ) -> Sequence[tuple[float, int | None, float | None]]:
         """Query-only current interval and S90 per input from the resident state.
 
         One forward pass per card, no state bytes across the bridge, GIL
